refactor: 使用 TypeScript 进行重构
