+++ packages/vue-to-dsl/package.json
@@ -9,11 +9,16 @@
   "main": "dist/tiny-engine-vue-to-dsl.cjs",
   "module": "dist/tiny-engine-vue-to-dsl.js",
   "types": "dist/index.d.ts",
+  "bin": {
+    "tiny-vue-to-dsl": "dist/cli.js"
+  },
   "files": [
     "dist"
   ],
   "scripts": {
-    "build": "vite build",
+    "build": "pnpm run build:types && vite build && pnpm run build:cli",
+    "build:types": "tsc -p tsconfig.json",
+    "build:cli": "tsc -p tsconfig.cli.json",
     "test": "vitest",
     "test:unit": "vitest run",
     "coverage": "vitest run --coverage",
@@ -31,15 +36,18 @@
   "license": "MIT",
   "homepage": "https://opentiny.design/tiny-engine",
   "dependencies": {
-    "@vue/compiler-sfc": "^3.4.15",
-    "@vue/compiler-dom": "^3.4.15",
     "@babel/parser": "^7.23.0",
     "@babel/traverse": "^7.23.0",
     "@babel/types": "^7.23.0",
+    "@vue/compiler-dom": "^3.4.15",
+    "@vue/compiler-sfc": "^3.4.15",
     "vue": "^3.4.15"
   },
   "devDependencies": {
+    "@types/node": "^20.11.24",
     "@vitest/coverage-v8": "^1.4.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.0",
     "vite": "^5.4.2",
     "vitest": "^1.4.0"
   }
