+++ packages/vue-to-dsl/src/index.ts
@@ -1,0 +1,7 @@
+import './index.d.ts'
+
+export { VueToDslConverter } from './converter'
+export { parseVueFile, parseSFC } from './parser'
+export { generateSchema, generateAppSchema } from './generator'
+export { parseTemplate, parseScript, parseStyle } from './parsers'
+export * from './types/index'
