+++ packages/vue-to-dsl/vite.config.js
@@ -3,12 +3,24 @@ import { defineConfig } from 'vite'
 export default defineConfig({
   build: {
     lib: {
-      entry: 'src/index.js',
+      entry: 'src/index.ts',
       name: 'TinyEngineVueToDsl',
+      formats: ['es', 'cjs'],
       fileName: (format) => `tiny-engine-vue-to-dsl.${format === 'es' ? 'js' : format}`
     },
     rollupOptions: {
-      external: ['vue', '@vue/compiler-sfc', '@vue/compiler-dom', '@babel/parser', '@babel/traverse', '@babel/types'],
+      external: [
+        'vue',
+        '@vue/compiler-sfc',
+        '@vue/compiler-dom',
+        '@babel/parser',
+        '@babel/traverse',
+        '@babel/types',
+        'node:fs',
+        'node:fs/promises',
+        'node:path',
+        'node:url'
+      ],
       output: {
         globals: {
           vue: 'Vue',
